+++ sources-doc/Changelog.md
@@ -1,4 +1,5 @@
-* 2021-Aug: Adding code and documentation for Diaphony (Mayur Patel, mayur.patel.rogers@gmail.com)
+* 2021-Dec: Adding Kronecker K21-2 and R2 sample generators
+* 2021-Aug: Adding code and documentation for Diaphony
 * 2020-Nov: Cleanup in the Generalized L2 Discrepancy code
 * 2020-Jan: Bugfix FFTW_INCLUDES headers
 * 2020-Jan: More Jittered samplers